Move location integrity tests into their own file
